Move url_key_capacity() & url_key_remaining() to services
