+++ tests/Unit/Helpers/UrlHlpTest.php
@@ -2,11 +2,21 @@
 
 namespace Tests\Unit;
 
-use Facades\App\Helpers\UrlHlp;
+use App\Services\UrlService;
 use Tests\TestCase;
 
 class UrlHlpTest extends TestCase
 {
+    protected $url;
+
+    public function setUp():void
+    {
+        parent::setUp();
+
+        $this->url = new UrlService();
+
+    }
+
     public function test_url_limit()
     {
         $this->assertSame(
@@ -40,18 +50,18 @@ class UrlHlpTest extends TestCase
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', 0);
         config()->set('urlhub.hash_size_2', 0);
-        $this->assertSame(0, UrlHlp::url_key_capacity());
+        $this->assertSame(0, $this->url->url_key_capacity());
 
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', 1);
         config()->set('urlhub.hash_size_2', 2);
-        $this->assertSame(12, UrlHlp::url_key_capacity()); // (3^1)+(3^2)
+        $this->assertSame(12, $this->url->url_key_capacity()); // (3^1)+(3^2)
 
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', 2);
         config()->set('urlhub.hash_size_2', 2);
         // $alphabet_length^$hash_size_1 or 3^2
-        $this->assertSame(9, UrlHlp::url_key_capacity());
+        $this->assertSame(9, $this->url->url_key_capacity());
     }
 
     public function test_url_key_capacity_input_negative()
@@ -59,17 +69,17 @@ class UrlHlpTest extends TestCase
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', 1);
         config()->set('urlhub.hash_size_2', -2);
-        $this->assertSame(3, UrlHlp::url_key_capacity());
+        $this->assertSame(3, $this->url->url_key_capacity());
 
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', -1);
         config()->set('urlhub.hash_size_2', 2);
-        $this->assertSame(0, UrlHlp::url_key_capacity());
+        $this->assertSame(0, $this->url->url_key_capacity());
 
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', -1);
         config()->set('urlhub.hash_size_2', -2);
-        $this->assertSame(0, UrlHlp::url_key_capacity());
+        $this->assertSame(0, $this->url->url_key_capacity());
     }
 
     public function test_url_key_capacity_input_number()
@@ -77,12 +87,12 @@ class UrlHlpTest extends TestCase
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', 2.7);
         config()->set('urlhub.hash_size_2', 3);
-        $this->assertSame(36, UrlHlp::url_key_capacity()); // (3^2)+(3^3)
+        $this->assertSame(36, $this->url->url_key_capacity()); // (3^2)+(3^3)
 
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', 2);
         config()->set('urlhub.hash_size_2', 3.7);
-        $this->assertSame(36, UrlHlp::url_key_capacity()); // (3^2)+(3^3)
+        $this->assertSame(36, $this->url->url_key_capacity()); // (3^2)+(3^3)
     }
 
     public function test_url_key_capacity_input_string()
@@ -90,17 +100,17 @@ class UrlHlpTest extends TestCase
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', 'string');
         config()->set('urlhub.hash_size_2', 2);
-        $this->assertSame(0, UrlHlp::url_key_capacity());
+        $this->assertSame(0, $this->url->url_key_capacity());
 
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', 2);
         config()->set('urlhub.hash_size_2', 'string');
         // $alphabet_length^$hash_size_1 or 3^2
-        $this->assertSame(9, UrlHlp::url_key_capacity());
+        $this->assertSame(9, $this->url->url_key_capacity());
 
         config()->set('urlhub.hash_alphabet', 'abc');
         config()->set('urlhub.hash_size_1', 'string');
         config()->set('urlhub.hash_size_2', 'string');
-        $this->assertSame(0, UrlHlp::url_key_capacity());
+        $this->assertSame(0, $this->url->url_key_capacity());
     }
 }
